refactor: 요청, 응답 클래스 위치 수정
